moved MirrorScanWindow to simplespectra
