+++ src/main/java/net/sf/mzmine/modules/visualization/spectra/simplespectra/mirrorspectra/MirrorScanWindow.java
@@ -1,4 +1,21 @@
-package net.sf.mzmine.modules.visualization.spectra.mirrorspectra;
+/*
+ * Copyright 2006-2019 The MZmine 2 Development Team
+ * 
+ * This file is part of MZmine 2.
+ * 
+ * MZmine 2 is free software; you can redistribute it and/or modify it under the terms of the GNU
+ * General Public License as published by the Free Software Foundation; either version 2 of the
+ * License, or (at your option) any later version.
+ * 
+ * MZmine 2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+ * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License along with MZmine 2; if not,
+ * write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
+ * USA
+ */
+package net.sf.mzmine.modules.visualization.spectra.simplespectra.mirrorspectra;
 
 import java.awt.BorderLayout;
 import javax.swing.JFrame;
@@ -13,6 +30,11 @@ import net.sf.mzmine.util.spectraldb.entry.DBEntryField;
 import net.sf.mzmine.util.spectraldb.entry.SpectralDBEntry;
 import net.sf.mzmine.util.spectraldb.entry.SpectralDBPeakIdentity;
 
+/**
+ * Creates a window with a mirror chart to compare to scans
+ * 
+ * @author Robin Schmid
+ */
 public class MirrorScanWindow extends JFrame {
 
   private JPanel contentPane;
